To add language functionality, functions from poetics were moved to class Poetry. Poetry will be abstract class with basic functionality, language specific parts will be declarated in inhereted classes
