+++ poetrytools/__init__.py
@@ -1,1 +1,2 @@
 from .poetics import *
+from .countsyl import *
